requirements.txt
certifi==2022.6.15.2
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.3
oauthlib==3.2.1
pandas==1.4.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.10.1
urllib3==1.26.12
